refactor(codex): dedupe startup client cleanup
